+++ package.json
@@ -8,22 +8,22 @@
     "test": "test"
   },
   "scripts": {
-    "test": "node --experimental-strip-types --test --test-concurrency=1 --experimental-test-coverage --test-force-exit test/**/*.test.ts",
-    "test:coverage": "mkdir -p coverage && node --experimental-strip-types --test --test-concurrency=1 --experimental-test-coverage --test-reporter=lcov --test-reporter-destination coverage/lcov.info --test-force-exit test/**/*.test.ts",
-    "start": "NODE_OPTIONS=--experimental-strip-types node src/server.ts",
+    "test": "node --test --test-concurrency=1 --experimental-test-coverage --test-force-exit test/**/*.test.ts",
+    "test:coverage": "mkdir -p coverage && node --test --test-concurrency=1 --experimental-test-coverage --test-reporter=lcov --test-reporter-destination=coverage/lcov.info --test-reporter=spec --test-reporter-destination=stdout --test-force-exit test/**/*.test.ts",
+    "start": "node --import ./src/instrument.ts src/server.ts",
     "typecheck": "tsc",
-    "dev": "node --watch src/server.ts",
-    "info": "node src/bin/hav-print-routes.ts",
+    "dev": "node --watch --import ./src/instrument.ts src/server.ts",
+    "info": "node --import ./src/instrument.ts src/bin/hav-print-routes.ts",
     "lint": "biome check --write src/",
     "lint:check": "biome check src/",
-    "hav:init-mongodb": "node --experimental-strip-types src/bin/hav-init-mongodb.ts",
-    "hav:update-subscription-length": "node --experimental-strip-types src/bin/hav-update-subscription-length.ts",
-    "hav:populate-queue": "node --experimental-strip-types src/bin/hav-populate-queue.ts",
-    "hav:send-queue": "node --experimental-strip-types src/bin/hav-send-queue.ts",
-    "hav:test-sms-sending": "node --experimental-strip-types src/bin/hav-test-sms-sending.ts",
-    "hav:test-email-templates": "node --experimental-strip-types src/bin/hav-test-email-templates.ts",
-    "hav:test-all-templates": "node --experimental-strip-types src/bin/hav-test-all-templates.ts",
-    "hav:run-dialogi-test-server": "node --experimental-strip-types src/bin/hav-dialogi-test-server.ts"
+    "hav:init-mongodb": "node --import ./src/instrument.ts src/bin/hav-init-mongodb.ts",
+    "hav:update-subscription-length": "node --import ./src/instrument.ts src/bin/hav-update-subscription-length.ts",
+    "hav:populate-queue": "node --import ./src/instrument.ts src/bin/hav-populate-queue.ts",
+    "hav:send-queue": "node --import ./src/instrument.ts src/bin/hav-send-queue.ts",
+    "hav:test-sms-sending": "node --import ./src/instrument.ts src/bin/hav-test-sms-sending.ts",
+    "hav:test-email-templates": "node --import ./src/instrument.ts src/bin/hav-test-email-templates.ts",
+    "hav:test-all-templates": "node --import ./src/instrument.ts src/bin/hav-test-all-templates.ts",
+    "hav:run-dialogi-test-server": "node --import ./src/instrument.ts src/bin/hav-dialogi-test-server.ts"
   },
   "keywords": [],
   "author": "",
@@ -33,7 +33,7 @@
     "@fastify/mongodb": "^9.0.2",
     "@fastify/sensible": "^6.0.3",
     "@fastify/type-provider-typebox": "^6.1.0",
-    "@immobiliarelabs/fastify-sentry": "^9.0.1",
+    "@sentry/node": "^10.12.0",
     "@sinclair/typebox": "^0.34.41",
     "fastify": "^5.6.2",
     "fastify-mailer": "^2.3.1",
