+++ src/app.ts
@@ -1,7 +1,7 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import AutoLoad, { type AutoloadPluginOptions } from '@fastify/autoload';
-import fastifySentry from '@immobiliarelabs/fastify-sentry';
+import * as Sentry from '@sentry/node';
 import type { FastifyPluginAsync, FastifyPluginOptions } from 'fastify';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
@@ -12,28 +12,9 @@ export interface AppOptions extends FastifyPluginOptions, Partial<AutoloadPlugin
 export const options: AppOptions = {};
 
 const app: FastifyPluginAsync<AppOptions> = async (fastify, opts) => {
-  fastify.register(fastifySentry, {
-    dsn: process.env.SENTRY_DSN,
-    beforeSend: (event) => {
-      if (!event?.request?.data) {
-        return event;
-      }
-
-      const data = JSON.parse(event.request.data);
-
-      if (!data.email) {
-        return event;
-      }
-
-      delete data.email;
-      event.request.data = JSON.stringify(data);
-
-      return event;
-    },
-    environment: process.env.ENVIRONMENT,
-    release: process.env.SENTRY_RELEASE ?? '',
-    setErrorHandler: true,
-  });
+  // Sentry is initialized via the preloaded src/instrument.ts; this wires up
+  // the Fastify error handler so uncaught route errors are reported.
+  Sentry.setupFastifyErrorHandler(fastify);
 
   fastify.register(AutoLoad, {
     dir: join(__dirname, 'plugins'),
