--- src/plugins/sentry.ts
@@ -1,7 +1,0 @@
-import type Sentry from '@sentry/node';
-
-declare module 'fastify' {
-  export interface FastifyInstance {
-    Sentry: typeof Sentry;
-  }
-}
